export in 3 files, valid checks integrated. some small tweeks might be in place later
