--- brickwall
@@ -1,5 +1,0 @@
-hello world
-ls
-grep
-hello world
-echo
